+++ src/main/java/com/notenoughmail/kubejs_tfc/item/JugItemBuilder.java
@@ -1,5 +1,6 @@
 package com.notenoughmail.kubejs_tfc.item;
 
+import com.notenoughmail.kubejs_tfc.util.BuilderRefs;
 import com.notenoughmail.kubejs_tfc.util.ResourceUtils;
 import dev.latvian.mods.kubejs.client.LangEventJS;
 import dev.latvian.mods.kubejs.generator.AssetJsonGenerator;
@@ -29,8 +30,8 @@ public class JugItemBuilder extends ItemBuilder {
         super(i);
         capacity = () -> 100;
         acceptableFluids = TFCTags.Fluids.USABLE_IN_JUG;
-        FluidContainerItemBuilder.colorList.add(this);
-        FluidContainerItemBuilder.dispenserList.add(this);
+        BuilderRefs.fluidContainerDispenser.add(this);
+        BuilderRefs.fluidContainerColor.add(this);
     }
 
     @Info("Sets the display name to use when the jug is filled")