+++ changelog.md
@@ -4,12 +4,14 @@
 - Add direct wrapper for TFC's `ItemStackProvider`
 - Fix double crops not having yield multipliers in their loot tables
 - Add convenience method for setting the dead model of torches
-- Add sapling block builder
-- Rename several methods in crop block block builders to prevent Rhino throwing a fit when used
+- Add TFC sapling & leaf block builders
+- Rename several methods in crop block builders to prevent Rhino throwing a fit when used
   - `#productItem(ResourceLocation)` -> `#existingProductItem(ResourceLocation)`
   - `#model(number,Consumer)` -> `#setModel(number,Consumer)`
   - `#model(Consumer)` -> `#setModel(Consumer)`
 - Allow `RecipeJS#getOriginalRecipeIngredients()` to work with TFC recipe types
+- Add ability to change custom crop growth and expiry rates
+- Add `always` nad `never` shortcuts for extended property settings that take `StatePredicate`s
 
 ### 1.3.1
 - Fix supports having broken connecting models by default